Create divisible_sum.cpp
You are given an array of n integers, ar=[ar[0],...,ar[n-1]] , and a positive integer, k. Find and print the number of (i,j) pairs where i<j and arr[i]+arr[j]  is divisible by k.

For example,  and . Our three pairs meeting the criteria are  and .

Function Description

Complete the divisibleSumPairs function in the editor below. It should return the integer count of pairs meeting the criteria.

divisibleSumPairs has the following parameter(s):

n: the integer length of array 
ar: an array of integers
k: the integer to divide the pair sum by