Create greatXOR.cpp
Given a long integer x, count the number of values of a satisfying the following conditions:
a xor x > x
0 < a < x
where  and  are long integers and xor is the bitwise XOR operator.

You are given q queries, and each query is in the form of a long integer denoting x. For each query, print the total number of values of a
satisfying the conditions above on a new line.

For example, you are given the value x=5. Condition 2 requires that a<x. The following tests are run:
1 ^ 5 = 4
2 ^ 5 = 7
3 ^ 5 = 6
4 ^ 5 = 1
We find that there are 2 values meeting the first condition: 2 and 3.