Create Student_Person.cpp
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed code for Person and a declaration for Student are provided for you in the editor. Observe that Student inherits all the properties of Person.

Complete the Student class by writing the following:

A Student class constructor, which has  parameters:
firstName - A string denoting the Person's first name.
        *   lastName - A string denoting the Person's last name.
        *   id - An integer denoting the Person's ID number.
        *   scores - An array of integers denoting the Person's test scores
A char calculate() method that calculates a Student object's average and returns the grade character representative of their calculated average:
 if(avg<=100 && avg>=90)
                return 'O';
            if(avg<90 && avg>=80)
                return 'E';
            if(avg<80 && avg>=70)
                return 'A';
            if(avg<70 && avg>=55)
                return 'P';
            if(avg<55 && avg>=40)
                return 'D';
            return 'T';