Create bookFine.cpp
Your local library needs your help! Given the expected and actual return dates for a library book, create a program that calculates the fine (if any). The fee structure is as follows:

If the book is returned on or before the expected return date, no fine will be charged (i.e.: 0 fine.)
If the book is returned after the expected return day but still within the same calendar month and year as the expected return date, fine=15 X (the number of days late).
If the book is returned after the expected return month but still within the same calendar year as the expected return date, the fine=500 X (the number of months late).
If the book is returned after the calendar year in which it was expected, there is a fixed fine of 10000.